tests/requirements.txt
psycopg2
mock
django-jsonfield
tblib
